+++ test/integration/012_deprecation_tests/dispatch-macros/cast.sql
@@ -1,7 +1,7 @@
 
 
 {% macro string_literal(s) -%}
-  {{ adapter.dispatch('string_literal', macro_namespace='test')(s) }}
+  {{ adapter.dispatch('string_literal', packages=['test'])(s) }}
 {%- endmacro %}
 
 {% macro default__string_literal(s) %}