+++ test/integration/043_custom_aliases_test/macros/macros.sql
@@ -9,7 +9,7 @@
 
 
 {% macro string_literal(s) -%}
-  {{ adapter.dispatch('string_literal', packages=['test'])(s) }}
+  {{ adapter.dispatch('string_literal', macro_namespace='test')(s) }}
 {%- endmacro %}
 
 {% macro default__string_literal(s) %}